Revert "Broken sketch for keeping seed at the global level"
This reverts commit 3f1287e90051b4e2b071fd43799dd21f45477a35.
