+++ packages/unmock-core/src/__tests__/backend/index.test.ts
@@ -16,10 +16,6 @@ describe("Node.js interceptor", () => {
         isWhitelisted: (_: string) => false,
         log: (_: string) => undefined,
         useInProduction: () => false,
-        runnerOptions: {
-          maxLoop: 20,
-          seed: 0,
-        },
       });
     });
