+++ packages/unmock-core/src/index.ts
@@ -1,7 +1,7 @@
 // Sinon for asserts and matchers
 import * as sinon from "sinon";
 import NodeBackend from "./backend";
-import { ILogger, IUnmockOptions, IUnmockPackage, IRunnerOptions } from "./interfaces";
+import { ILogger, IUnmockOptions, IUnmockPackage } from "./interfaces";
 import WinstonLogger from "./loggers/winston-logger";
 import runner from "./runner";
 import { AllowedHosts, BooleanSetting } from "./settings";
@@ -11,14 +11,8 @@ export * from "./types";
 export { sinon };
 export { default as dsl } from "./service/state/transformers";
 
-const defaultRunnerOptions: IRunnerOptions = {
-  maxLoop: 20,
-  seed: 0,
-};
-
 export class UnmockPackage implements IUnmockPackage {
   public allowedHosts: AllowedHosts;
-  public runnerOptions: Partial<IRunnerOptions> = {};
   public flaky: BooleanSetting;
   public useInProduction: BooleanSetting;
   protected readonly backend: NodeBackend;
@@ -44,10 +38,6 @@ export class UnmockPackage implements IUnmockPackage {
       isWhitelisted: (url: string) => this.allowedHosts.isWhitelisted(url),
       log: (message: string) => this.logger.log(message),
       flaky: () => this.flaky.get(),
-      runnerOptions: {
-        ...defaultRunnerOptions,
-        ...this.runnerOptions,
-      },
     };
     this.backend.initialize(opts);
     return this;
