+++ packages/unmock-core/src/interfaces.ts
@@ -33,13 +33,7 @@ export interface IListener {
   }): void;
 }
 
-export interface IRunnerOptions {
-  maxLoop: number;
-  seed: number;
-}
-
 export interface IUnmockOptions extends ILogger {
-  runnerOptions: IRunnerOptions;
   useInProduction(): boolean;
   isWhitelisted(url: string): boolean;
   flaky(): boolean;
@@ -47,7 +41,6 @@ export interface IUnmockOptions extends ILogger {
 
 export interface IUnmockPackage {
   allowedHosts: AllowedHosts;
-  runnerOptions: Partial<IRunnerOptions>;
   services: ServiceStoreType;
   on(): IUnmockPackage;
   init(): IUnmockPackage;
